+++ src/onegov.ballot/HISTORY.rst
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
-- Inlcudes absolute majority in the election export.
+- Adds more information to the election export.
   [msom]
 
 0.4.2 (2016-03-17)